+++ backend/src/main/java/hanglog/trip/dto/response/TripLedgerResponse.java
@@ -6,9 +6,9 @@ import hanglog.city.dto.response.CityResponse;
 import hanglog.currency.domain.Currency;
 import hanglog.expense.domain.Amount;
 import hanglog.expense.domain.CategoryExpense;
-import hanglog.expense.domain.DayLogExpense;
 import hanglog.expense.dto.response.CategoryExpenseResponse;
 import hanglog.expense.dto.response.ExchangeRateResponse;
+import hanglog.trip.domain.DayLogExpense;
 import hanglog.trip.domain.Trip;
 import java.math.BigDecimal;
 import java.time.LocalDate;
