+++ backend/src/test/java/hanglog/community/presentation/CommunityControllerTest.java
@@ -32,11 +32,11 @@ import hanglog.community.dto.response.CommunityTripResponse;
 import hanglog.community.dto.response.RecommendTripListResponse;
 import hanglog.community.service.CommunityService;
 import hanglog.expense.domain.CategoryExpense;
-import hanglog.expense.domain.DayLogExpense;
+import hanglog.global.ControllerTest;
+import hanglog.trip.domain.DayLogExpense;
+import hanglog.trip.dto.response.TripDetailResponse;
 import hanglog.trip.dto.response.TripLedgerResponse;
 import hanglog.trip.service.LedgerService;
-import hanglog.global.ControllerTest;
-import hanglog.trip.dto.response.TripDetailResponse;
 import jakarta.servlet.http.Cookie;
 import java.time.LocalDateTime;
 import java.util.List;
