+++ backend/src/test/java/hanglog/expense/service/LedgerServiceTest.java
@@ -26,10 +26,10 @@ import hanglog.city.domain.repository.CityRepository;
 import hanglog.currency.domain.repository.CurrencyRepository;
 import hanglog.expense.domain.Amount;
 import hanglog.expense.domain.CategoryExpense;
-import hanglog.expense.domain.DayLogExpense;
-import hanglog.trip.dto.response.TripLedgerResponse;
 import hanglog.expense.fixture.ExchangeableExpenseFixture.ExchangeableExpense;
+import hanglog.trip.domain.DayLogExpense;
 import hanglog.trip.domain.repository.TripRepository;
+import hanglog.trip.dto.response.TripLedgerResponse;
 import hanglog.trip.service.LedgerService;
 import java.math.BigDecimal;
 import java.util.Arrays;
