+++ backend/src/test/java/hanglog/trip/presentation/TripControllerTest.java
@@ -41,15 +41,15 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import hanglog.auth.domain.MemberTokens;
 import hanglog.city.domain.City;
 import hanglog.expense.domain.CategoryExpense;
-import hanglog.expense.domain.DayLogExpense;
-import hanglog.trip.dto.request.SharedStatusRequest;
-import hanglog.trip.dto.response.SharedCodeResponse;
-import hanglog.trip.dto.response.TripLedgerResponse;
 import hanglog.global.ControllerTest;
+import hanglog.trip.domain.DayLogExpense;
 import hanglog.trip.dto.request.PublishedStatusRequest;
+import hanglog.trip.dto.request.SharedStatusRequest;
 import hanglog.trip.dto.request.TripCreateRequest;
 import hanglog.trip.dto.request.TripUpdateRequest;
+import hanglog.trip.dto.response.SharedCodeResponse;
 import hanglog.trip.dto.response.TripDetailResponse;
+import hanglog.trip.dto.response.TripLedgerResponse;
 import hanglog.trip.dto.response.TripResponse;
 import hanglog.trip.service.LedgerService;
 import hanglog.trip.service.TripService;
